requirements.txt
absl-py==2.1.0
addict==2.4.0
asttokens==2.4.1
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.1
ConfigArgParse==1.7
contourpy==1.2.0
cuda-python==12.4.0
cupy-cuda12x==13.1.0
cycler==0.12.1
dash==2.15.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
executing==2.0.1
fastjsonschema==2.19.1
fastrlock==0.8.2
filelock==3.13.1
Flask==3.0.2
fonttools==4.48.1
fsspec==2024.2.0
future==0.18.3
grpcio==1.60.1
h5py==3.10.0
huggingface-hub==0.20.3
hydra-core==0.11.3
idna==3.6
importlib-metadata==7.0.1
ipython==8.21.0
ipywidgets==8.1.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_core==5.7.1
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
KNN-CUDA @ https://github.com/unlimblue/KNN_CUDA/releases/download/0.2/KNN_CUDA-0.2-py3-none-any.whl#sha256=809a43da2eed950c65ba15e355e8cdbf975b6a948dced1f02414f26aba586cd0
llvmlite==0.42.0
lmdb==1.4.1
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mpmath==1.3.0
msgpack==1.0.7
msgpack-numpy==0.4.8
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
numba==0.59.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.133
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
nvitop==1.3.2
omegaconf==1.4.1
open3d==0.18.0
packaging==23.2
pandas==2.2.0
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
plotly==5.18.0
pointnet2-ops @ file:///home/wsui/projects/Pointnet2_PyTorch/pointnet2_ops_lib
prompt-toolkit==3.0.43
protobuf==4.25.2
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
Pygments==2.17.2
pylibraft-cu12==24.4.0
pyparsing==3.1.1
pyquaternion==0.9.9
python-dateutil==2.8.2
pytorch-lightning==0.7.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
retrying==1.3.4
rmm-cu12==24.4.0
rpds-py==0.18.0
safetensors==0.4.2
scikit-learn==1.4.0
scipy==1.12.0
six==1.16.0
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.16.1
tensorboard-data-server==0.7.2
termcolor==2.4.0
tf-keras==2.15.0
threadpoolctl==3.3.0
tifffile==2024.2.12
timm==0.9.12
torch==2.2.0
torchaudio==2.2.0
torchsummary==1.5.1
torchvision==0.17.0
tqdm==4.66.2
traitlets==5.14.1
triton==2.2.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
wcwidth==0.2.13
Werkzeug==3.0.1
widgetsnbextension==4.0.10
zipp==3.17.0
